Chante terms: Use ENVIRONMENT instead of CONTEXT
<xire-context> is currently defined to hold xire macro bindings and it
is designed to hold other information for further development.

But the purpose of <xire-context> is like a Scheme environment and
another kind of context will be introduced to xire.  The latter context
will represent the context of xire compilation such as top-level,
statement and expression.  So it's very confusing if there are two
"context"s.

Fix the problem by using term "environment" instead of "context" (which
was defined in terms <xire-context>).