xire-compile: Repat macro expansion recursively
For ease of writing macros which are expanded into forms with the same
macros, change xire-compile to repeat macro expansion until a given form
cannot be expanded.

For example, we have to write a recursive macro as follows with old
versions:

	(define-xire-macro :statement (recursion-test form context)
	  (match form
	    [(_) (xire-compile '(recursion-test "A") context)]
	    [(_ a) (xire-compile `(recursion-test "B" ,a) context)]
	    [(_ a b) (xire-compile `("result" ,a ,b) context)]))

Now we can write the same macro as follows with the new version:

	(define-xire-macro :statement (recursion-test form context)
	  (match form
	    [(_) '(recursion-test "A")]
	    [(_ a) `(recursion-test "B" ,a)]
	    [(_ a b) `("result" ,a ,b)]))

New version of xire-compile accepts both styles, though there is
a little overhead in the latter style.