Treat unknown expression as a functioni call
When compiling (f x y z) in expression context, the old version raises
error because "f" is not registered as an expression macro.  Though this
treatment is not wrong, it's hard to deal with user-defined functions.

So that the new version treats such expression as a function call to
simplify the implementation of xire compiler.