Refuse unknown macro in non-expression context
Most of works in early stage of xire development are to write an
imaginary xire script and to update xirec to compile the script.
Currently, unknown macro in non-expression context is evaluated as
a Scheme expression.  But this treatment is not useful for the above
style.

Change the compiler to refuse unknown macro in non-expression context.