Add a way to embed Scheme script in xire script
The old way to embed Scheme script in xire script is not intuitive and
it has a problem for further development.  So that it is already
removed.  But it is necessary to embed Scheme script for some reason.

Add (scheme SCHEME-SCRIPT ...) notation for the top-level context to
embed Scheme script.