Rename functions to create an iform
Functions to create an iform are named with prefix "make-".  It might be
readable without any knowledge.  However, "make-" is somewhat long to
type and its length makes code disordering.

Rename the functions with new prefix "$".  The "$" prefix resembles tags
of iforms.  Though the functions and the tags have the same spellings,
symbols in Gauche are case-sensitive by default, so that it is not
a problem to use such naming.